requirements.txt

fastapi
uvicorn
web3
httpx
pydantic
numpy
sklearn.cluster
sklearn.preprocessing